pk-file: make .load aware of pkgdatadir
This commit makes the .load command to look in pkgdatadir
(PREFIX/share/poke/) for the specified file, if it is not found
relative to the current working directory.

In the future, we will want to support a language-level `load' that
will look in a user-defined, program-initialized, poke_load_path.  But
for now, .load must suffice.

2019-10-01  Jose E. Marchesi  <jemarch@gnu.org>

	* src/pk-file.c (pk_cmd_load_file): Try to load the specified file
	relative to the current working directory, then look in datadir.