pickles: move non-loadable pickles to a future/ subdirectory
Many of the pickles in pickles/ do not load because they use some
facility that is not yet implemented, while others use weird syntax
and concepts that predate the current form of the language; I wrote
these to get an idea of what I wanted/needed.

To avoid people getting confused (especially in this early stage where
we do not have a manual) I am moving all the pickles that cannot be
loaded to pickles/future, and changing the build system to not install
them.

As we support more features, we can move/adapt the files back from
`future' into the pickles directory.

2019-10-01  Jose E. Marchesi  <jemarch@gnu.org>

	* pickles/Makefile.am (dist_pkgdata_DATA): Distribute elf.pk and
	ctf.pk only.
	* pickles/future/bmp.pk: Moved from pickles/
	* pickles/future/tar.pk: Likewise.
	* pickles/future/swf.pk: Likewise.
	* pickles/future/pdp10.pk: Likewise.
	* pickles/future/mp3.pk: Likewise.
	* pickles/future/leb128.pk: Likewise.
	* pickles/future/jpeg.pk: Likewise.
	* pickles/future/id3v2.pk: Likewise.
	* pickles/future/id3v1.pk: Likewise.
	* pickles/future/gzip.pk: Likewise.
	* pickles/future/flv.pk: Likewise.
	* pickles/future/elf-future.pk: Likewise.
	* pickles/future/dwarf.pk: Likewise.
	* pickles/future/deflate.pk: Likewise.
	* pickles/future/bson.pk: Likewise.