pkl: allocate pkl-asm data structures using the GC
The macro-assembler structures contain PVM programs, which are
allocated with malloc (inside Jitter) and in turn contain PVM values
which are allocated with the GC malloc.  This easily leads to
situations where there are no roots routing to to these PVM values and
they get collected.

This patch makes the asm data structures to be allocated with
GC_malloc.  This fixes the current problems that led to the
deactivation of the GC.

GC is also activated again :)

2019-10-17  Jose E. Marchesi  <jemarch@gnu.org>

	* src/pkl-asm.c (pkl_asm_pushlevel): Use pvm_alloc instead of
	malloc.
	(pkl_asm_new): Likewise.
	(pkl_asm_poplevel): Do not free memory.
	(pkl_asm_finish): Likewise.
	* src/pvm-alloc.c (pvm_alloc_initialize): Enable garbage
	collection.