pk,pkl: partial support for error-on-warning
This patch adds a new command `set error-on-warning (yes|no)' to
pk-cmd.  It also fixes the command infrastructure to accept commands
including `-' characters in their names.

Finally, it adds a configurable flag to the PKL compiler.  However,
this flag is not yet used by pkl_warning.  The reason is that we
should make the compiler passes aware of the current compiler, passing
it as an argument to pkl_do_pass.  Also, convenient wrappers shall be
provided in PKL_PASS_EMIT_ERROR and PKL_PASS_EMIT_WARNING.

2019-10-19  Jose E. Marchesi  <jemarch@gnu.org>

	* src/pkl.c (struct pkl_compiler): New field error_on_warning.
	(pkl_error_on_warning): New function.
	(pkl_set_error_on_warning): Likewise.
	* src/pkl.h: Prototypes for pkl_error_on_warning and
	pkl_set_error_on_warning.
	* src/pk-cmd.c (pk_cmd_exec_1): Support `-' characters in command
	names.
	* src/pk-set.c (set_error_on_warning_cmd): New variable.
	(pk_cmd_set_error_on_warning): New function.
	(set_cmds): Add set_error_on_warning_cmd.